requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
requests==2.31.0
ollama==0.1.6
python-dotenv==1.0.0
